requirements.txt
arrow==0.15.4
Pillow==6.2.1
requests==2.22.0
beautifulsoup4==4.8.1
